config_manager: move 4 core file pairs into src/config_manager/
Phase A of the configuration-manager extraction: pure relocation, no
semantic change. The four pairs moved are:

  src/deptran/raft/config_manager.{h,cc}   -> src/config_manager/
  src/deptran/raft/cluster_config.{h,cc}   -> src/config_manager/
  src/deptran/raft/config_watcher.{h,cc}   -> src/config_manager/
  src/deptran/config_store.{h,cc}          -> src/config_manager/

These all sit "above" the storage / consensus layer and conceptually
belong to the master-shard / configuration-manager component described
in docs/mako-book.md §3. Co-locating them makes the boundary visible
in the layout and lets a follow-up commit introduce an abstract
ReplicatedDB interface so the CM can be unit-tested against an
in-memory fake instead of pulling in real Raft + RocksDB.

External callers (config_node_init, config_service, raft/test.cc, the
existing tests/config_*_test.cc, and src/rrr/tests/sharding_startup_test.cc)
have their #include paths updated to "config_manager/<file>.h".

CMake: src/config_manager/*.cc is added to DEPTRAN_CORE_SRC so the
files keep landing in txlog_core. The three RAFT_BORROW_SRC exclude
lines for config_manager.cc / cluster_config.cc / config_watcher.cc
are dropped since those files no longer live under src/deptran/raft/.
replicated_db.cc stays in raft/ (still excluded from borrow checking
because of <rocksdb/c.h>); it's the implementation behind the
abstraction Phase B will introduce.
