+++ src/config_manager/config_manager.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "replicated_db.h"
+#include "deptran/raft/replicated_db.h"
 #include <string>
 #include <vector>
