+++ tests/config_service_test.cc
@@ -7,7 +7,7 @@
 
 #include <gtest/gtest.h>
 #include "../src/deptran/config_service.h"
-#include "../src/deptran/config_store.h"
+#include "../src/config_manager/config_store.h"
 #include "../src/deptran/config_schema.h"
 #include "../src/deptran/sharding_policy.h"
 #include "../src/deptran/sharding_policy_builder.h"
